# Models/StoreDto.cs
namespace AdminDashboard.Api.Models;

public class StoreDto
{
    public int Id { get; set; }
    public string Code { get; set; } = "";
    public string Name { get; set; } = "";
    public string? Address { get; set; }
    public string? Phone { get; set; }
    public string Status { get; set; } = "Active";
}

public class StoreCreateUpdateDto
{
    public string Code { get; set; } = "";
    public string Name { get; set; } = "";
    public string? Address { get; set; }
    public string? Phone { get; set; }
    public string Status { get; set; } = "Active";
}