# openapi.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  description: API cho hệ thống đặt hàng đa NCC (Multi-Supplier Order). Chi tiết xem API-FE-LIEN-KET.md
  version: 1.0.0
servers:
  - url: /api
    description: Base path API
...
paths:
...
  /suppliers/{id}:
    put:
      tags: [Suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplierInput' }
      security: [{ Bearer: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supplier' }
    delete:
      tags: [Suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '204': { description: Đã xóa }
...
  /products:
    get:
      tags: [Products]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: supplierId
          in: query
          schema: { type: integer }
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [Active, Inactive] }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [Products]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      security: [{ Bearer: [] }]
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
  /products/{id}:
    put:
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
    delete:
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '204': { description: Đã xóa }
  /stores:
    get:
      tags: [Stores]
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Store' }
    post:
      tags: [Stores]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreInput' }
      security: [{ Bearer: [] }]
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Store' }
  /stores/{id}:
    put:
      tags: [Stores]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreInput' }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Store' }
    delete:
      tags: [Stores]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '204': { description: Đã xóa }
...
  /orders/{id}:
    get:
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
...
  /supplier-orders:
    get:
      tags: [SupplierOrders]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrderSupplier' }
  /supplier-orders/{id}:
    get:
      tags: [SupplierOrders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderSupplier' }
  /supplier-orders/{id}/status:
    patch:
      tags: [SupplierOrders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [Confirmed, Rejected, Partial, Delivering, Delivered]
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderSupplier' }
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token trả về từ POST /auth/login
  schemas:
...
    Supplier:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        contact: { type: string }
        email: { type: string }
        address: { type: string }
        status: { type: string }
    SupplierInput:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        contact: { type: string }
        email: { type: string }
        address: { type: string }
        status: { type: string }
...
    Product:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        supplierId: { type: integer }
        categoryId: { type: integer }
        unit: { type: string }
        price: { type: number }
        status: { type: string }
    ProductInput:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        supplierId: { type: integer }
        categoryId: { type: integer }
        unit: { type: string }
        price: { type: number }
        status: { type: string }
    Store:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        status: { type: string }
    StoreInput:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        status: { type: string }
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unit: { type: string }
        price: { type: number }
        discountAmount: { type: number }
        taxPercent: { type: number }
        taxAmount: { type: number }
...
    ReceiveImage:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [received, invoice] }
        imageUrl: { type: string }
        fileName: { type: string }
    OrderSupplier:
      type: object
      properties:
        id: { type: integer }
        orderId: { type: integer }
        supplierId: { type: integer }
        supplierName: { type: string }
        status: { type: string }
        expectedDeliveryDate: { type: string, format: date }
        actualDeliveryDate: { type: string, format: date, nullable: true }
        confirmDate: { type: string, format: date, nullable: true }
        note: { type: string }
        totalAmount: { type: number }
        taxAmount: { type: number }
        paymentStatus: { type: string }
        paidAmount: { type: number }
        paidDate: { type: string, format: date, nullable: true }
        receivedBy: { type: integer, nullable: true }
        receivedDate: { type: string, format: date-time, nullable: true }
        orderItems:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        receiveImages:
          type: array
          items: { $ref: '#/components/schemas/ReceiveImage' }
...
    Order:
      type: object
      properties:
        id: { type: integer }
        storeId: { type: integer }
        storeName: { type: string }
        status: { type: string }
        orderDate: { type: string, format: date }
        expectedDeliveryDate: { type: string, format: date, nullable: true }
        totalItemCount: { type: integer }
        totalAmount: { type: number }
        createdBy: { type: integer }
        createdDate: { type: string, format: date-time }
        orderSuppliers:
          type: array
          items: { $ref: '#/components/schemas/OrderSupplier' }
...